Moved all schemas to be part of each operation
