+++ docetl/operations/base.py
@@ -8,7 +8,15 @@ from typing import Dict, List, Optional, Tuple
 from docetl.operations.utils import APIWrapper
 from rich.console import Console
 from rich.status import Status
+import jsonschema
+from pydantic import BaseModel
 
+# FIXME: This should probably live in some utils module?
+class classproperty(object):
+    def __init__(self, f):
+        self.f = f
+    def __get__(self, obj, owner):
+        return self.f(owner)
 
 class BaseOperation(ABC):
     def __init__(
@@ -45,6 +53,16 @@ class BaseOperation(ABC):
         )
         self.syntax_check()
 
+    # This must be overridden in a subclass
+    class schema(BaseModel):
+        name: str
+        type: str
+
+    @classproperty
+    def json_schema(cls):
+        assert hasattr(cls.schema, "model_json_schema"), "Programming error: %s.schema must be a pydantic object but is a %s" % (cls, type(cls.schema))
+        return cls.schema.model_json_schema()
+    
     @abstractmethod
     def execute(self, input_data: List[Dict]) -> Tuple[List[Dict], float]:
         """
@@ -73,7 +91,7 @@ class BaseOperation(ABC):
         Raises:
             ValueError: If the configuration is invalid.
         """
-        pass
+        jsonschema.validate(instance = self.config, schema = self.json_schema)
 
     def gleaning_check(self) -> None:
         """
