+++ docetl/operations/gather.py
@@ -1,4 +1,4 @@
-from typing import Any, Dict, List, Tuple
+from typing import Any, Dict, List, Tuple, Optional
 
 from docetl.operations.base import BaseOperation
 
@@ -15,6 +15,14 @@ class GatherOperation(BaseOperation):
     5. Return results containing the rendered chunks with added context, including information about skipped characters and headers.
     """
 
+    class schema(BaseOperation.schema):
+        type: str = "gather"
+        content_key: str
+        doc_id_key: str
+        order_key: str
+        peripheral_chunks: Dict[str, Any]
+        doc_header_key: Optional[str] = None
+    
     def __init__(self, *args: Any, **kwargs: Any) -> None:
         """
         Initialize the GatherOperation.
